safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEColorMatrixElement-svgdom-values-prop-expected.txt

CONSOLE MESSAGE: line 186: TypeError: 'undefined' is not an object (evaluating 'matrixElement.values.baseVal')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'values' property of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL SVGNumberListToString(matrixElement.values.baseVal) should be 0.9 0.9 0.9 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 0.33 0.33 0.33 0 0 . Threw exception TypeError: 'undefined' is not an object (evaluating 'matrixElement.values.baseVal')
FAIL Math.round(saturateElement.values.baseVal.getItem(0).value * 1000) should be 500. Threw exception TypeError: 'undefined' is not an object (evaluating 'saturateElement.values.baseVal')
FAIL Math.round(hueRotateElement.values.baseVal.getItem(0).value * 1000) should be 10000. Threw exception TypeError: 'undefined' is not an object (evaluating 'hueRotateElement.values.baseVal')
